+++ CHANGELOG.md
@@ -15,6 +15,7 @@ guidelines from here: https://github.com/olivierlacan/keep-a-changelog
 - PLOS Biology name is no longer hard coded in letter templates
 - Fixed issue where roles couldn't be assigned in the admin page in some journals
 - Fixed ad-hoc email component to expand properly
+- Notifications when FTP uploads fail are now being emailed out to the Journal Staff Admins for associated papers
 ### Security
 
 ## [1.27.0] - {2016-10-6}